+++ delphi-frontend/src/test/java/au/com/integradev/delphi/antlr/GrammarTest.java
@@ -242,4 +242,9 @@ class GrammarTest {
   void testUnusualWhitespace() {
     parseFile("UnusualWhitespace.pas");
   }
+
+  @Test
+  void testUnusualIntegerLiterals() {
+    parseFile("UnusualIntegerLiterals.pas");
+  }
 }