+++ delphi-frontend/src/test/resources/au/com/integradev/delphi/grammar/UnusualIntegerLiterals.pas
@@ -1,0 +1,23 @@
+unit UnusualIntegerLiterals;
+
+interface
+
+const
+  HexWithoutDigits = $;
+  BinaryWithoutDigits = %;
+  HexStartingWithUnderscore = $_;
+  BinaryStartingWithUnderscore = %_;
+  HexStartingWithManyUnderscores = $__;
+  BinaryStartingWithManyUnderscores = %__;
+
+  CharHexStartingWithUnderscore = #$_;
+  CharBinaryStartingWithUnderscore = #%_;
+  CharHexStartingWithManyUnderscores = #$__;
+  CharBinaryStartingWithManyUnderscores = #%__;
+
+  CharDecimalStartingWithUnderscore = #_;
+  CharDecimalStartingWithManyUnderscores = #__;
+
+implementation
+
+end.